% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latex_support.R
\name{add_latex_support}
\alias{add_latex_support}
\title{Add support for latex to quarto file}
\usage{
add_latex_support(filename)
}
\description{
Add support for latex to quarto file
}
\keyword{internal}
