% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latex_support.R
\name{add_to_quarto_file}
\alias{add_to_quarto_file}
\alias{add_to_rmarkdown_file}
\title{Add webexercises 'YAML' metadata}
\usage{
add_to_quarto_file(filename)

add_to_rmarkdown_file(filename)
}
\arguments{
\item{filename}{The name of the file to edit.}
}
\value{
No return value, called for side effects.
}
\description{
Adds the necessary 'YAML' metadata to an existing 'Quarto' or 'Rmarkdown'
file.
}
