% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{answer}
\alias{answer}
\title{LearnR-style answer}
\usage{
answer(text, correct = FALSE, message = NULL)
}
\arguments{
\item{text}{The answer text}

\item{correct}{Whether this as a/the correct answer}

\item{message}{(Not used)}
}
\value{
named vectot
}
\description{
LearnR-style answer
}
