% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesomebox.R
\name{box_args}
\alias{box_args}
\title{Print an awesomebox}
\usage{
box_args(
  vrulecolor = "white",
  hrule = c("\\\\abLongLine", "\\\\abShortLine", ""),
  title = "",
  vrulewidth = "0pt",
  icon = "Question",
  iconcolor = "black"
)
}
\description{
Print an awesomebox
}
\keyword{internal}
