% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cli_msg}
\alias{cli_msg}
\title{Wraps cli messages for a beautiful command line user interface;
Use argument names:
"!" = "This is a warning"
"v" = "This is a green checkmark"
"x" = "This is a red cross"}
\usage{
cli_msg(...)
}
\description{
Wraps cli messages for a beautiful command line user interface;
Use argument names:
"!" = "This is a warning"
"v" = "This is a green checkmark"
"x" = "This is a red cross"
}
\keyword{internal}
