% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webexercises_fns.R
\name{mcq}
\alias{mcq}
\title{Create a multiple-choice question}
\usage{
mcq(opts)
}
\arguments{
\item{opts}{Vector of alternatives. The correct answer is the
element(s) of this vector named 'answer'.}
}
\value{
A character string with HTML code to generate a pull-down
  menu.
}
\description{
Create a multiple-choice question
}
\details{
Writes html code that creates an option box widget, with one or
  more correct answers. Call this function inline in an RMarkdown document.
  See the Web Exercises RMarkdown template for further examples.
}
\examples{
# How many planets orbit closer to the sun than the Earth?
mcq(c(1, answer = 2, 3))

# Which actor played Luke Skywalker in the movie Star Wars?
mcq(c("Alec Guinness", answer = "Mark Hamill", "Harrison Ford"))
}
