% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{question}
\alias{question}
\title{LearnR-style question}
\usage{
question(
  text,
  ...,
  type = c("auto", "single", "multiple", "learnr_radio", "learnr_checkbox",
    "learnr_text"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", text, "<br/><br/><br/>"),
  submit_button = "Submit Answer",
  try_again_button = "Try Again",
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)
}
\arguments{
\item{text}{The question text}

\item{...}{Answers (see webexercises::answer)}

\item{type}{The input type (only single for now)}

\item{correct}{(not used yet)}

\item{incorrect}{(not used yet)}

\item{try_again}{(not used yet)}

\item{message}{(not used yet)}

\item{post_message}{(not used yet)}

\item{loading}{(not used yet)}

\item{submit_button}{(not used yet)}

\item{try_again_button}{(not used yet)}

\item{allow_retry}{(not used yet)}

\item{random_answer_order}{Whether to randomise the order of the answers}

\item{options}{(not used yet)}
}
\value{
html for the webexercises question
}
\description{
LearnR-style question
}
\examples{
question("Is R fun?",
         answer("Yes", TRUE),
         answer("No"),
         random_answer_order = TRUE)
}
