% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_to_quiz.R
\name{quiz_from_cvs}
\alias{quiz_from_cvs}
\title{Read a quiz from a csv spreadsheet and pass it to webexercises::quiz}
\usage{
quiz_from_cvs(filename)
}
\description{
Read a quiz from a csv spreadsheet and pass it to webexercises::quiz
}
\keyword{internal}
