% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latex_support.R
\name{rmd_webex_support}
\alias{rmd_webex_support}
\title{Add webexercises support}
\usage{
rmd_webex_support()
}
\value{
Character string of class `HTML`.
}
\description{
This convenience function can be called within an 'Rmarkdown'
document to add the 'css' and 'JavaScript' code required for webexercises'
html output.
}
\examples{
invisible(rmd_webex_support())
}
\seealso{
\code{\link[htmltools]{HTML}}
}
