% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_questions.R
\name{round2}
\alias{round2}
\title{Round up from .5}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{digits}{Integer indicating the number of decimal places (`round`) or significant digits (`signif`) to be used.}
}
\value{
A vector of rounded numeric values.
}
\description{
Round up from .5
}
\details{
Implements rounding using the "round up from .5" rule,
  which is more conventional than the "round to even" rule
  implemented by R's built-in \code{\link{round}} function.
}
\examples{
round2(c(2, 2.5))

# compare to:
round(c(2, 2.5))
}
