% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webexercises_fns.R
\name{style_widgets}
\alias{style_widgets}
\title{Change webexercises widget style}
\usage{
style_widgets(incorrect = "orchid3", correct = "lightgreen")
}
\arguments{
\item{incorrect}{The colour of the widgets when the answer is incorrect (defaults to orchid3).}

\item{correct}{The colour of the widgets when the correct answer
not filled in (defaults to lightgreen).}
}
\value{
A character string containing HTML code to change the CSS
  style values for widgets.
}
\description{
Change default colours using R colour names (see `colours()`) or hexadecimal notation
}
\details{
Call this function inline in an RMarkdown document to
  change the default and correct colours using any valid CSS colour
  word (e.g., red, rgb(255,0,0), hsl(0, 100%, 50%) or #FF0000).
}
\examples{
style_widgets("yellow", "purple")
}
