% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webexercises_fns.R
\name{torf}
\alias{torf}
\title{Create a true-or-false question}
\usage{
torf(answer)
}
\arguments{
\item{answer}{Logical value TRUE or FALSE, corresponding to the correct answer.}
}
\value{
A character string with HTML code to generate a pull-down
  menu with elements TRUE and FALSE.
}
\description{
Create a true-or-false question
}
\details{
Writes html code that creates an option box widget with TRUE or FALSE as alternatives. Call this function inline in an RMarkdown document. See the Web Exercises RMarkdown template for further examples.
}
\examples{
# True or False? 2 + 2 = 4
torf(TRUE)

# True or False? The month of April has 31 days.
torf(FALSE)
}
