% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webexercises_default.R
\name{webexercises_default2}
\alias{webexercises_default2}
\title{Create webexercises document in bookdown::html_document2 style}
\usage{
webexercises_default2(css, ...)
}
\arguments{
\item{css}{Path of css file to add.}

\item{...}{Additional function arguments to pass to
\code{\link[bookdown]{html_document2}}.}
}
\value{
R Markdown output format to pass to 'render'.
}
\description{
This function wraps \code{bookdown::html_document2} to configure
compilation to embed the default webexercises CSS and JavaScript files in
the resulting HTML.
}
\details{
Call this function as the \code{output_format} argument
  for the \code{\link[rmarkdown]{render}} function when compiling
  HTML documents from RMarkdown source.
}
\examples{
# copy the webexercises 'R Markdown' template to a temporary file
\dontrun{
my_rmd <- tempfile(fileext = ".Rmd")
rmarkdown::draft(my_rmd, "webexercises", "webexercises")

# compile it
rmarkdown::render(my_rmd, webexercises::webexercises_default2())

# view the result
browseURL(sub("\\\\.Rmd$", ".html", my_rmd))
}
}
\seealso{
\code{\link[rmarkdown]{render}}, \code{\link[bookdown]{html_document2}}
}
