% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{with_cli_try}
\alias{with_cli_try}
\title{Wraps cli for a beautiful command line user interface which indicates if
`code` passes or fails.}
\usage{
with_cli_try(msg, code, ...)
}
\description{
Wraps cli for a beautiful command line user interface which indicates if
`code` passes or fails.
}
\keyword{internal}
