% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_quarto.R
\name{add_to_quarto}
\alias{add_to_quarto}
\title{Add webexercises helper files to quarto}
\usage{
add_to_quarto(
  quarto_dir = ".",
  include_dir = "include",
  output_format = c("html")
)
}
\arguments{
\item{quarto_dir}{The base directory for your quarto project}

\item{include_dir}{The directory where you want to put the css and
js files (defaults to "include")}

\item{output_format}{The format you want to add
webexercises to (only html for now)}
}
\value{
No return value, called for side effects.
}
\description{
Adds the necessary helper files to an existing quarto project and
edits the _quarto.yml file accordingly. A demo file for webexercises
will be added and optionally rendered.
}
