% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{create_quarto_doc}
\alias{create_quarto_doc}
\title{Create a quarto document with webexercise}
\usage{
create_quarto_doc(name = "Untitled", open = interactive())
}
\arguments{
\item{name}{Name of the new document}

\item{open}{Whether to open the document in RStudio}
}
\value{
The file path to the document
}
\description{
Creates a new directory with the file name and copies in a demo qmd file and the necessary helper files.
}
