% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{quiz}
\alias{quiz}
\title{LearnR-style quiz}
\usage{
quiz(..., caption = "Quiz")
}
\arguments{
\item{...}{one or more questions (see webexercises::question)}

\item{caption}{Optional quiz caption}
}
\value{
html
}
\description{
LearnR-style quiz
}
