% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webexercises_fns.R
\name{textbox}
\alias{textbox}
\title{Add a textbox}
\usage{
textbox(
  rows = 10,
  cols = 100,
  border_size = 2,
  border_color = "#000000",
  text = NULL
)
}
\arguments{
\item{rows}{Number of rows.}

\item{cols}{Height or columns.}

\item{border_size}{Size of the border.}

\item{border_color}{Color of the border.}

\item{text}{Prefilled text. Default to `getOption("webexercices_textbox_text")`.}
}
\value{
Html text.
}
\description{
Add a textbox
}
